tinker/examples/mnist/requirements.txt
tensorflow-gpu>=2.2.0
matplotlib>=3.3.1
